+++ frontend/tests/pages/newsletter/confirmation.test.tsx
@@ -1,6 +1,6 @@
 import { render, waitFor } from "@testing-library/react";
 import { axe } from "jest-axe";
-import NewsletterConfirmation from "src/pages/newsletter-confirmation";
+import NewsletterConfirmation from "src/pages/newsletter/confirmation";
 
 describe("Newsletter", () => {
   it("passes accessibility scan", async () => {
