+++ frontend/tests/pages/newsletter/index.test.tsx
